+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 📚 Documentation
 
+- update deps ([04e99a9](git@github.com:monako97/expression-language-editor/commit/04e99a91a1d0d499dda1f2cca480c5331e53fce3))
 - update deps ([f82bbdc](git@github.com:monako97/expression-language-editor/commit/f82bbdcbaa02a36a2b965643e959f02e56777c4f))
 
 ## v0.0.2